Split XML helpers into typed extensions
